refactor: 공고 리스트 컴포넌트 리네임 및 디렉토리 구조 변경
