--- src/app/components/CardMockData.ts
@@ -1,76 +1,0 @@
-const CardMockData = [
-  {
-    image: '/images/mock/ex1.png',
-    title: '물방개맛집식당',
-    date: '2025-02-31',
-    hours: '15:00-18:00 (3시간)',
-    location: '서울시 강남구',
-    price: '15,000원',
-    discount: '기존 시급보다 50%',
-  },
-  {
-    image: '/images/mock/ex2.png',
-    title: '맛집북카페',
-    date: '2025-02-12',
-    hours: '14:00-18:00 (4시간)',
-    location: '서울시 성수구',
-    price: '12,000원',
-    discount: '기존 시급보다 30%',
-  },
-  {
-    image: '/images/mock/ex3.png',
-    title: '도토리식당',
-    date: '2025-01-02',
-    hours: '15:00-18:00 (3시간)',
-    location: '서울시 강남구',
-    price: '10,000원',
-    discount: '기존 시급보다 10%',
-  },
-  {
-    image: '/images/mock/ex1.png',
-    title: '수리수리밥줘',
-    date: '2025-04-31',
-    hours: '13:00-18:00 (5시간)',
-    location: '서울시 강남구',
-    price: '40,000원',
-    discount: '기존 시급보다 60%',
-  },
-  {
-    image: '/images/mock/ex1.png',
-    title: '최고야식당',
-    date: '2025-01-31',
-    hours: '13:00-17:00 (4시간)',
-    location: '서울시 은평구',
-    price: '20,000원',
-    discount: '기존 시급보다 30%',
-  },
-  {
-    image: '/images/mock/ex3.png',
-    title: '큰맘줄게식당',
-    date: '2025-03-23',
-    hours: '15:00-18:00 (3시간)',
-    location: '서울시 용산구',
-    price: '17,000원',
-    discount: '기존 시급보다 30%',
-  },
-  {
-    image: '/images/mock/ex2.png',
-    title: '사랑카페',
-    date: '2025-01-08',
-    hours: '13:00-19:00 (6시간)',
-    location: '서울시 서대문구',
-    price: '30,000원',
-    discount: '기존 시급보다 20%',
-  },
-  {
-    image: '/images/mock/ex1.png',
-    title: '행복버거',
-    date: '2025-04-31',
-    hours: '14:00-18:00 (4시간)',
-    location: '서울시 노원구',
-    price: '40,000원',
-    discount: '기존 시급보다 100%',
-  },
-];
-
-export default CardMockData;
