+++ src/app/components/notice/DetailedFilter.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState } from 'react';
-import { LOCATION_LIST } from '../constants/location';
+import { LOCATION_LIST } from '../../constants/location';
 
 const inputClass =
   'h-14 w-full rounded-md border border-gray-30 px-4 py-1 text-gray-black focus:outline-none focus:ring';
@@ -54,7 +54,7 @@ const DetailedFilter: React.FC = () => {
 
           <div className="mb-4">
             <h3 className="mb-2 text-base">위치</h3>
-            <div className="custom-scrollbar grid max-h-48 grid-cols-2 gap-2 overflow-auto rounded-md border p-3">
+            <div className="grid max-h-48 grid-cols-2 gap-2 overflow-auto rounded-md border p-3 custom-scrollbar">
               {LOCATION_LIST.map((location) => (
                 <button
                   key={location}
@@ -106,7 +106,7 @@ const DetailedFilter: React.FC = () => {
                   type="number"
                   value={amount}
                   onChange={(e) => setAmount(e.target.value)}
-                  className={`${inputClass} no-spinner pr-10`}
+                  className={`${inputClass} pr-10 no-spinner`}
                   placeholder="입력"
                 />
                 <span className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-black">
